src/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-JWT-Extended==4.4.4
flask-sqlalchemy==3.0.2
flask-marshmallow==0.14.0

flake8==4.0.1
black==22.6.0
isort==5.10.1

toml==0.10.2
python-dateutil==2.8.2
python-dotenv==0.20.0
bcrypt==4.0.1

gunicorn==20.1.0
